requirements.txt
selenium==4.35.0
requests==2.32.4
beautifulsoup4==4.13.4
pillow==11.3.0
lxml==6.0.0
pytesseract==0.3.10
PyAutoGUI==0.9.54
python-dotenv==1.1.1
webdriver-manager==4.0.2

python-docx==1.2.0
sqlalchemy==2.0.41
pymysql==1.1.2

typing_extensions==4.14.1
pyodbc==5.2.0
uvicorn
fastapi
apscheduler
httpx

